simulador/requirements.txt
paho-mqtt==2.0.0
python-dotenv==2.0.0
